+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 * Use constants instead of hardcoded value for query limits. Refs UICIRC-724.
 * Provide read-only access to loan policies. Refs UICIRC-693.
 * Prevent deletion of policies if they're in use. Refs UICIRC-669.
+* Add RTL/Jest testing for `FeeFineNoticesSection` component in `NoticePolicy/components/EditSections/FeeFineNoticesSection`. Refs UICIRC-637.
+
 
 ## [6.0.0] (https://github.com/folio-org/ui-circulation/tree/v6.0.0) (2021-09-30)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v5.1.1...v6.0.0)